+++ lib/iso8601/errors.rb
@@ -4,23 +4,33 @@ module ISO8601
   ##
   # Contains all ISO8601-specific errors.
   module Errors
+
+    class StandardError < ::StandardError
+    end
     ##
     # Raised when the given pattern doesn't fit as ISO 8601 parser.
-    class UnknownPattern < ::StandardError
+    class UnknownPattern < StandardError
       def initialize(pattern)
         super("The pattern “#{pattern}” is not allowed in this implementation of ISO8601.")
       end
     end
     ##
+    # Raised when the given pattern contains an invalid fraction.
+    class InvalidFractions < StandardError
+      def initialize(pattern)
+        super("Fractions are only allowed in the last component")
+      end
+    end
+    ##
     # Raised when the given date is valid but out of range.
-    class RangeError < ::StandardError
+    class RangeError < StandardError
       def initialize(pattern)
         super("“#{pattern}” is out of range")
       end
     end
     ##
     # Raise when the base is not suitable.
-    class DurationBaseError < ::StandardError
+    class DurationBaseError < StandardError
       def initialize(duration)
         super("Wrong base for #{duration} duration.")
       end